data_crawler/requirements.txt
PyMySQL>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
yfinance>=0.2.30
ccxt>=4.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
lxml>=4.9.0
pytz>=2023.3
pandas>=2.0.0
